0.1.3: Added declarative events, param.sanitizer property, defvalue property.
Declarative Events: Can now declare events in svidget markup directly
i.e. onchange="myFunc".
Param sanitizer property: Can wire a func to call when a param is set to
change, validate the value
Param "valuechange" renamed to "set". The event is triggered when a
value is changed on a param.
New defvalue property for Param and ActionParam.